Fix AwesomeWM configuration for version 4.0
This is a major release which breaks lots of API and breaks other. The
main seems to be regarding Vicious. It is not compatible with the
changes to the widget API.

The main workaround left is about the `set_vertical` for both memory and
alsa widgets. This function has no effect in this version forcing to
using a wibox.container.rotate and flip the widget's width and height.

A lot of unhandled errors and warnings are getting printed at
~/.xsession-errors. The keyboard layout widget does not seem to work as
well.