Replace Google Chrome by Chromium
I had Google Chrome installed before because it was the only broser
supported by Netflix. Which is no longer true.

I replace it by Chromium because I still may need it for mobile app
development for example. Given this work related necessity, I also move
it to the workspace role.

Closes #18