Share Axiom (mass storage) using samba on BURN-E
This will make possible for me to share files between my machines, watch
content on TV and even remote access.

We still need to create an user account and set a samba password for it.

The actual storage is yet to arrive as I have to fix the PSU to support
the two disks.