Fix samba case sensitivity
The default configuration has case sensitivity disabled causing some of
the files and folders I'm backing up to raise prompts of what to do
about merging folders and replacing files with the same name.

This is one of the things Windows makes out life harder. As I do not
plan to use this share with Windows machines, I'm here enabling the case
sensitivity.